Adds method to generate an Etch-a-Sketch grid
Commit adds the buildEAS(dimension) function. The function adds divs
to the DOM to create a square Etch-a-Sketch grid. The size of the
grid is set by the dimension parameter - the function creates a
dimension x dimension grid using the etchRow and etchBox classes as
descendents of the #etchFrame div.