Makes etch grid dimensions always 513px x 513px
Later instructions in the Odin Project project specified that the
size of the grid should not change regardless of the number of boxes
in the Etch-a-Sketch grid. This commit makes it so the total grid is
always 513px x 513px, adding border around the grid to fill necessary
space.

This commit also does a little cleanup in buildEAS, using Math.floor
to calculate the etch box div size, rather than a more complicated
comparison operation using Math.round.