Makes etch-a-sketch shade 10% per hover
Extra credit assignment for the Project

Previously, on each hover the etchBox div would change from white to
black. Now, on each hover, the color gets progressively darker, by
~10% of the way from white to black each hover. This is accomplished
by changing the onHover event to call a new function, shade(color)
which takes in a string representing a color of the format
'rgb(red, green, blue)' and returns a string in HEX-RGB format that
is a shade of grey 26 RGB-units darker than the red value of the
input. This commit also adds a helper function, dec2hex(dec) which
takes an input integer and returns its Hex value as a string (without
the leading #).